wip2

+++ src/scripts/synthetic_graph_generator.py
@@ -66,7 +66,7 @@ def _generate_single_grid_graph(size):
     return nx.relabel_nodes(grid, mapping)
 
 
-def single_graph_generator(graph_family: str, no_of_vertices: int, ratio: tuple[int, int], **kwargs):
+def single_graph_generator(graph_family: str, no_of_vertices: int, ratio: tuple[float, float], **kwargs):
     filename = ""
 
     match graph_family:
